requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML utilities
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.2.0

# Data processing - use specific versions with pre-built wheels
h5py>=3.8.0; platform_system != "Windows"
h5py==3.11.0; platform_system == "Windows"
tqdm>=4.65.0
PyYAML>=6.0

# CUDA support
# For RTX 4060: Install PyTorch with CUDA 11.8 or 12.1
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
